Refactor Data package, Remove Month object
