+++ app/src/main/java/edu/sjsu/posturize/posturize/PostureData/PostureMeasurement.java
@@ -1,4 +1,4 @@
-package edu.sjsu.posturize.posturize.SexyData;
+package edu.sjsu.posturize.posturize.PostureData;
 
 import java.util.Date;
