--- app/src/main/java/edu/sjsu/posturize/posturize/SexyData/MonthlyPosture.java
@@ -1,17 +1,0 @@
-package edu.sjsu.posturize.posturize.SexyData;
-
-/**
- * Created by Matthew on 8/28/2017.
- *
- * TODO:
- * DONE - Create object to track months activity
- * Create method to add new DailyPosture objects to month
- * Create method to grab specific day of month
- */
-
-/**
- * MonthlyPosture contains an array of DailyPosture objects to manage the months measurements
- */
-public class MonthlyPosture {
-    private DailyPosture[] monthlyData;
-}
